@@ -1,6 +1,13 @@
 module github.com/banzaicloud/imps
 
-go 1.18
+go 1.24.4
+
+replace (
+	k8s.io/api => k8s.io/api v0.31.0
+	k8s.io/apimachinery => k8s.io/apimachinery v0.31.0
+	k8s.io/client-go => k8s.io/client-go v0.31.0
+	sigs.k8s.io/controller-runtime => sigs.k8s.io/controller-runtime v0.19.0
+)
 
 require (
 	emperror.dev/emperror v0.33.0
