@@ -91,7 +91,7 @@ type ImagePullSecretStatus struct {
 	Status                       ReconciliationStatus `json:"status,omitempty"`
 	Reason                       string               `json:"reason,omitempty"`
 	LastSuccessfulReconciliation *metav1.Time         `json:"lastSuccessfulReconciliation,omitempty"`
-	ValiditySeconds              int32                `json:"validitySeconds,omitempty"`
+	ValiditySeconds              int64                `json:"validitySeconds,omitempty"`
 	ManagedNamespaces            []string             `json:"managedNamespaces,omitempty"`
 	SourceSecretStatus           map[string]string    `json:"sourceSecretStatus,omitempty"`
 }
